Share cache with different Accept headers matching with entity tags
git-svn-id: http://repo.aduna-software.org/svn/org.openrdf/alibaba/trunk@9528 3fbd4d3e-0f96-47c4-bab7-89622e02a19e

+++ metadata-server/src/main/java/org/openrdf/server/metadata/cache/CachedResponse.java
@@ -247,28 +247,20 @@ public class CachedResponse {
 			InterruptedException {
 		Lock lock = locker.getWriteLock();
 		try {
-			this.status = store.getStatus();
-			String statusText = store.getStatusText();
-			this.statusText = statusText == null ? "" : statusText;
 			setHeaders(store);
-			File tmp = store.getMessageBody();
-			if (body.exists()) {
-				body.delete();
-			}
-			if (tmp != null) {
-				tmp.renameTo(body);
+			if (store.isModified()) {
+				this.status = store.getStatus();
+				String statusText = store.getStatusText();
+				this.statusText = statusText == null ? "" : statusText;
+				File tmp = store.getMessageBody();
+				if (body.exists()) {
+					body.delete();
+				}
+				if (tmp != null) {
+					tmp.renameTo(body);
+				}
 			}
 			writeHeaders(head);
-		} finally {
-			lock.release();
-		}
-	}
-
-	public void setResponseHeaders(FileResponse store)
-			throws InterruptedException {
-		Lock lock = locker.getWriteLock();
-		try {
-			setHeaders(store);
 		} finally {
 			lock.release();
 		}
